+++ plugins/source/aws/resources/services/iam/groups.go
@@ -1,6 +1,10 @@
 package iam
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -33,8 +37,42 @@ func Groups() *schema.Table {
 		},
 
 		Relations: []*schema.Table{
-			GroupPolicies(),
+			groupPolicies(),
 			groupLastAccessedDetails(),
 		},
 	}
+}
+
+func fetchIamGroups(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	var config iam.ListGroupsInput
+	svc := meta.(*client.Client).Services().Iam
+	for {
+		response, err := svc.ListGroups(ctx, &config)
+		if err != nil {
+			return err
+		}
+		res <- response.Groups
+		if aws.ToString(response.Marker) == "" {
+			break
+		}
+		config.Marker = response.Marker
+	}
+	return nil
+}
+
+func resolveIamGroupPolicies(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	r := resource.Item.(types.Group)
+	svc := meta.(*client.Client).Services().Iam
+	config := iam.ListAttachedGroupPoliciesInput{
+		GroupName: r.GroupName,
+	}
+	response, err := svc.ListAttachedGroupPolicies(ctx, &config)
+	if err != nil {
+		return err
+	}
+	policyMap := map[string]*string{}
+	for _, p := range response.AttachedPolicies {
+		policyMap[*p.PolicyArn] = p.PolicyName
+	}
+	return resource.Set(c.Name, policyMap)
 }
